1.Add new modules:pcie/pvt/wdt/timer. 2.Add ddr for fpga. 3.Complete cru test 4.Add i2c passive mode function. 5.Add DDR FE/CU DMA function.

+++ src/driver/src/acu_ddrc_x8_ddr3_init.c
@@ -2,7 +2,7 @@
 
 void ddrc_x8_ddr3_init(int chs)
 {
-    int i = 0;
+    uint8_t i = 0;
 
     /////////////////////////////////////////////
     // step 1 : Program the DWC_ddr_umctl2 registers
